Restructure some of the cs files
