+++ src/Sweetener.Logging/ContextualLoggers/Templates/ILogEntryTemplate{T}.cs
