+++ src/Sweetener.Logging/Loggers/Templates/ILogEntryTemplate.cs
